requirements.txt
dj-database-url==0.4.1
Django==1.9.7
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.1
PyYAML==3.12
whitenoise==2.0.6
